requirements.txt
streamlit
transformers>=4.30.0
torch>=2.0.0
torchvision
torchaudio
sentence-transformers>=2.2.2
faiss-gpu
pandas
numpy
datasets
requests
python-dotenv
beautifulsoup4
langchain
langchain-community
langchain-huggingface
rank_bm25
accelerate
bitsandbytes
